move CompressPointCloudOptions out of draco.rs into remote_access
CompressPointCloudOptions is an algorithm-agnostic selector (a non-exhaustive
enum meant to grow other variants), but it lived inside draco.rs -- the module
we just reframed as the pure Draco mechanism. The pure encoder never uses it
(compress_point_cloud takes DracoEncodeOptions directly); its only consumers
are the remote-access policy layer, and its draco_options() accessor was
already remote-access-gated. So it was a remote-access-layer type squatting in
the algorithm module, with the dependency direction inverted.

Move it next to the PointCloudCompression trait in
remote_access/point_cloud_compression.rs and re-export it, leaving draco.rs as
just the encoding mechanism. Public path becomes
foxglove::remote_access::CompressPointCloudOptions. Consumers repointed; the
redundant cfg on draco_options() dropped.
